Modified the way trade adjustments are handled
When an order in the trading strategy is filled, adjustments were
made in two steps, cancel trades that needed adjustment and post
new adjusted trades. Now the first trade, which is never an adjustmnet,
will be posted, and all trades following will be canceled and adjusted
one at a time. This should aliviate error logging when later canceling
trades that never posted due to being post-only and market conditions
changing while canceling all of the trades that need adjusting. Step
towards #48