Handling of post-only rejection
Errors were caused by unhandled rejected post-only that would trade if not
for post-only flag when there was a later attempt to cancel. Much of this
was mitigate by the previous commit, but it still was possible that the first
trade may indeed face this issue if there was a quick swing in the opposite
direction of the mathced trade causing the adjustement. Closes #48