Isolated async socket manager from synchronous code using threads
By isolating the synchronous book manager from being called within the
async socket manager, the socket manager is able to continue to receive
messages from the socket and not be blocked by the posting of orders
within the book manager. Refactored book to work with database. Added
abstract class for a number of classes to make testing easier and
replacing mocking.

Using abstract classes noop implementations replace the calls to the
exchange api outside of the exchange api unit tests. By creating an
abstract class for the Trader, implementing new exchanges will be much
easier to implement.

closes #23
closes #35
works towards #24