Flask App implementation
By adding Flask to the project, I was able to make it easier to control
multiple trading pairs with multiple strategies with a nohup command.
Prior to this, strategies had to be started one by one, and due to rate
limiting done by the exchange api, which could lead to rejected api
requests. Now one process can be started and run indefinitely while also
allowing for the additions and removals of trading strategies as
desired.

Additions include
* Adding an Operations interface for the app to use to conrtol all
trading.
* Refactored Socket to subscribe and unsubscribe to pairs as needed.
* Socket know adds messages to a Queue that is consumed rather than
  containing the Reader within the class.
* Reader is now in its own thread.
* ThreadHandler handles all threads for all strategies in order to
  coordinate exchange api requests to avoid exceeding rate limits.
* Added in memory Mysql database to speed up testing of Flask routes
  and separate test data from the database (More to do in other tests.
* Utilized flask-migrate to manage database migrations. Also overwrote
  default scripts to keep a production and test production database in
  sync so that UAT environment trading on test exchange can have its
  own data separate from production data, but with matching schema.

fixes #55, #26, #19
towards #54, #24